+++ monad-eth-testutil/examples/txgen/shared/ecmul.rs
@@ -92,7 +92,7 @@ impl ECMul {
         max_fee_per_gas: u128,
         chain_id: u64,
         gas_limit: Option<u64>,
-        priority_fee: Option<u64>,
+        priority_fee: Option<u128>,
     ) -> TxEnvelope {
         let input = IECMul::performzksyncECMulsCall {
             iterations: U256::from(200),
@@ -104,7 +104,7 @@ impl ECMul {
             nonce: sender.nonce,
             gas_limit: gas_limit.unwrap_or(2_000_000), // 2M default, override with --set-tx-gas-limit
             max_fee_per_gas,
-            max_priority_fee_per_gas: priority_fee.unwrap_or(0) as u128, // 0 default, override with --priority-fee
+            max_priority_fee_per_gas: priority_fee.unwrap_or(0), // 0 default, override with --priority-fee
             to: TxKind::Call(self.addr),
             value: U256::ZERO,
             access_list: Default::default(),
